:sparkles: [#66] major 디렉토리 생성 및 파일 옮기기
