+++ src/main/java/Umc/replendar/major/controller/lectureController.java
@@ -1,0 +1,11 @@
+package Umc.replendar.major.controller;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/api/lectures")
+@RequiredArgsConstructor
+public class lectureController {
+}
