+++ src/main/java/Umc/replendar/user/service/MajorService.java
@@ -2,10 +2,10 @@ package Umc.replendar.user.service;
 
 import Umc.replendar.user.dto.req.MajorDtoReq;
 import Umc.replendar.user.dto.res.MajorDtoRes;
-import Umc.replendar.user.entity.Major;
-import Umc.replendar.user.entity.School;
-import Umc.replendar.user.repository.MajorRepository;
-import Umc.replendar.user.repository.SchoolRepository;
+import Umc.replendar.major.entity.Major;
+import Umc.replendar.major.entity.School;
+import Umc.replendar.major.repository.MajorRepository;
+import Umc.replendar.major.repository.SchoolRepository;
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
