+++ src/main/java/Umc/replendar/user/service/UserService.java
@@ -6,12 +6,13 @@ import Umc.replendar.common.security.JwtTokenProvider;
 import Umc.replendar.friend.repository.FriendRepository;
 import Umc.replendar.global.util.AmazonS3Util;
 import Umc.replendar.global.util.CookieUtil;
+import Umc.replendar.major.entity.Major;
 import Umc.replendar.user.converter.UserConverter;
 import Umc.replendar.user.dto.req.UserDtoReq;
 import Umc.replendar.user.dto.res.KakaoUserInfoResponseDto;
 import Umc.replendar.user.dto.res.UserDtoRes;
 import Umc.replendar.user.entity.*;
-import Umc.replendar.user.repository.MajorRepository;
+import Umc.replendar.major.repository.MajorRepository;
 import Umc.replendar.user.repository.UserRepository;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
